# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Asteria Drive MVP API
  version: 0.1.0-mvp
  summary: Control-plane API for the isolated Asteria Drive MVP
  description: |
    Asteria Drive coordinates tenant-scoped namespace operations, S3 multipart
    uploads, download authorizations, and recycle-bin operations. File bytes do
    not pass through this API.

    Business operations use either an OIDC/OAuth2 Bearer JWT (production) or a
    server-configured trusted-development bearer token (isolated development).
    OIDC requests select an active tenant with X-Tenant-ID; tenant membership
    and role are read from PostgreSQL rather than trusted from JWT claims.
servers:
  - url: http://127.0.0.1:8080
    description: Isolated local development
...
security:
  - oidcBearer: []
  - trustedDevBearer: []
...
paths:
...
  /api/v1/invitations/accept:
    post:
      tags: [Governance]
      operationId: acceptTenantInvitation
      summary: Accept an invitation as a verified external identity
      description: The bearer token is verified without requiring an existing tenant membership. Its issuer and subject must exactly match the invitation. The invitation token is supplied only in the JSON body.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptInvitationRequest"
      responses:
        "200":
          description: Invitation accepted and membership created, or the original accepted result replayed.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcceptInvitationEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/RequestTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/DependencyUnavailable"
  /api/v1/tenant:
    parameters:
      - $ref: "#/components/parameters/TenantSelector"
    get:
      tags: [Tenant]
      operationId: getCurrentTenant
      summary: Get the authenticated tenant and root directory
      description: |
        In trusted-dev the tenant is fixed by the bearer-token mapping. In OIDC
        mode the request must include X-Tenant-ID for an active membership.
      responses:
        "200":
          description: Current tenant.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantEnvelope"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/DependencyUnavailable"
  /api/v1/tenant/members:
    parameters:
      - $ref: "#/components/parameters/TenantSelector"
    get:
      tags: [Tenant]
      operationId: listTenantMembers
      summary: List tenant members
      description: Lists active and suspended members for owner/admin principals.
      parameters:
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageCursor"
      responses:
        "200":
          description: Stable keyset page ordered by principal ID.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/DependencyUnavailable"
  /api/v1/tenant/members/{principal_id}:
    parameters:
      - name: principal_id
        in: path
        required: true
        description: Tenant-scoped principal UUID.
        schema:
          type: string
          format: uuid
      - $ref: "#/components/parameters/TenantSelector"
    patch:
      tags: [Tenant]
      operationId: updateTenantMember
      summary: Change a tenant member role or status
      description: Owners may update any member. Admins cannot modify owners or grant the owner role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemberRequest"
      responses:
        "200":
          description: Updated member.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/DependencyUnavailable"
    delete:
      tags: [Tenant]
      operationId: deleteTenantMember
      summary: Delete a tenant membership
      responses:
        "204":
          description: Membership deleted.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/DependencyUnavailable"
  /api/v1/tenant/invitations:
    parameters:
      - $ref: "#/components/parameters/TenantSelector"
    get:
      tags: [Governance]
      operationId: listTenantInvitations
      summary: List tenant invitations
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/InvitationStatus"
      responses:
        "200":
          description: Tenant invitations.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationListEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/DependencyUnavailable"
    post:
      tags: [Governance]
      operationId: createTenantInvitation
      summary: Create a tenant invitation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvitationRequest"
      responses:
        "201":
          description: Invitation created; token is returned only by this response.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateInvitationEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/RequestTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/DependencyUnavailable"
  /api/v1/tenant/invitations/{invitation_id}/revoke:
    parameters:
      - name: invitation_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: "#/components/parameters/TenantSelector"
    post:
      tags: [Governance]
      operationId: revokeTenantInvitation
      summary: Revoke a pending tenant invitation
      responses:
        "200":
          description: Invitation revoked.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/DependencyUnavailable"
  /api/v1/tenant/groups:
    parameters:
      - $ref: "#/components/parameters/TenantSelector"
    get:
      tags: [Governance]
      operationId: listTenantGroups
      summary: List tenant groups
      responses:
        "200":
          description: Groups ordered by normalized name.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupListEnvelope"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Governance]
      operationId: createTenantGroup
      summary: Create a tenant group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupMutationRequest"
      responses:
        "201":
          description: Group created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/RequestTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/tenant/groups/{group_id}:
    parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: "#/components/parameters/TenantSelector"
    patch:
      tags: [Governance]
      operationId: updateTenantGroup
      summary: Rename a tenant group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupMutationRequest"
      responses:
        "200":
          description: Updated group.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Governance]
      operationId: deleteTenantGroup
      summary: Delete a tenant group and its memberships and ACL grants
      responses:
        "204":
          description: Group deleted.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/tenant/groups/{group_id}/members:
    parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: "#/components/parameters/TenantSelector"
    get:
      tags: [Governance]
      operationId: listTenantGroupMembers
      summary: List group members
      responses:
        "200":
          description: Group members ordered by principal ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberListEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/tenant/groups/{group_id}/members/{principal_id}:
    parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: principal_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: "#/components/parameters/TenantSelector"
    put:
      tags: [Governance]
      operationId: addTenantGroupMember
      summary: Idempotently add a tenant member to a group
      responses:
        "204":
          description: Group membership exists.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Governance]
      operationId: removeTenantGroupMember
      summary: Idempotently remove a member from a group
      responses:
        "204":
          description: Group membership is absent.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/tenant/audit-events:
    parameters:
      - $ref: "#/components/parameters/TenantSelector"
    get:
      tags: [Audit]
      operationId: listTenantAuditEvents
      summary: Page append-only tenant audit events
      parameters:
        - $ref: "#/components/parameters/AuditAfterSequence"
        - $ref: "#/components/parameters/AuditFrom"
        - $ref: "#/components/parameters/AuditUntil"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: Audit events ordered by sequence.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditPageEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/DependencyUnavailable"
  /api/v1/tenant/audit-events/export:
    parameters:
      - $ref: "#/components/parameters/TenantSelector"
    get:
      tags: [Audit]
      operationId: exportTenantAuditEvents
      summary: Export a bounded UTC range as NDJSON
      parameters:
        - $ref: "#/components/parameters/AuditFromRequired"
        - $ref: "#/components/parameters/AuditUntilRequired"
      responses:
        "200":
          description: At most 10,000 audit events as one JSON object per line.
          content:
            application/x-ndjson:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/DependencyUnavailable"
  /api/v1/directories:
    parameters:
      - $ref: "#/components/parameters/TenantSelector"
    post:
      tags: [Namespace]
      operationId: createDirectory
      summary: Create a directory
      description: Matching retries with the same Idempotency-Key replay the original directory. Reusing the key with a different request returns idempotency_conflict.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDirectoryRequest"
      responses:
        "201":
          description: Directory created.
          headers:
            Location:
              description: Relative URL of the created directory.
              schema:
                type: string
              example: /api/v1/directories/17da10b8-70d1-4f83-8d37-e127d110ef75
            ETag:
              $ref: "#/components/headers/ETag"
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            Idempotency-Replayed:
              $ref: "#/components/headers/IdempotencyReplayed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/RequestTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/DependencyUnavailable"
  /api/v1/directories/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceID"
      - $ref: "#/components/parameters/TenantSelector"
    get:
      tags: [Namespace]
      operationId: getDirectory
      summary: Get an active directory
      responses:
        "200":
          description: Directory metadata.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeEnvelope"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/DependencyUnavailable"
  /api/v1/directories/{id}/children:
    parameters:
      - $ref: "#/components/parameters/ResourceID"
      - $ref: "#/components/parameters/TenantSelector"
    get:
      tags: [Namespace]
      operationId: listDirectoryChildren
      summary: List direct active children
      parameters:
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageCursor"
      responses:
        "200":
          description: Stable keyset page ordered by normalized name and ID.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodePage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/DependencyUnavailable"
...
  /api/v1/nodes/{id}/acl/{entry_id}:
    parameters:
      - $ref: "#/components/parameters/ResourceID"
      - name: entry_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: "#/components/parameters/TenantSelector"
    delete:
      tags: [Governance]
      operationId: deleteNodeACL
      summary: Delete a direct ACL grant
      responses:
        "204":
          description: Grant deleted.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    oidcBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        OIDC/OAuth2 access token validated as a Resource Server. The token's
        issuer, signature, audience, expiry, and subject are verified; tenant
        and role come from PostgreSQL membership.
    trustedDevBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque trusted-development token
      description: |
        A high-entropy server-configured token mapped to exactly one tenant and
        principal. This mode is only for isolated development and acceptance.
...
  headers:
    RequestID:
      description: Validated incoming request ID or a server-generated request ID.
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: '^[A-Za-z0-9_.-]+$'
...
  schemas:
...
    InvitationStatus:
      type: string
      enum: [pending, accepted, revoked, expired]
...
    AcceptInvitationRequest:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token:
          type: string
          pattern: '^[0-9a-f]{64}$'
...